# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Open Card Control Plane API
  version: 1.1
  description: >-
    REST JSON and persistent SSE control-plane endpoints. Product state remains owned by the
    control plane; provider and agent endpoints are versioned contracts. API clients may send
    an exact Open-Card-API-Version header or an Open-Card-API-Min-Version and
    Open-Card-API-Max-Version range. All version headers are optional for backwards
    compatibility: clients that omit them negotiate the legacy 1.0 contract. Requests with
    no common supported version receive 426 Upgrade Required.
servers:
  - url: /
paths:
...
  /api/v1/applications:
    parameters:
      - $ref: '#/components/parameters/APIVersion'
      - $ref: '#/components/parameters/APIMinVersion'
      - $ref: '#/components/parameters/APIMaxVersion'
    get:
      operationId: listApplications
      responses:
        '200':
          description: Applications
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
    post:
      operationId: createApplication
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Created
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApplicationResponse'
        '400':
          description: Invalid request
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
        '409':
          description: Idempotency key was reused with different input
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
  /api/v1/applications/{applicationId}:
    parameters:
      - $ref: '#/components/parameters/ApplicationID'
      - $ref: '#/components/parameters/APIVersion'
      - $ref: '#/components/parameters/APIMinVersion'
      - $ref: '#/components/parameters/APIMaxVersion'
    get:
      operationId: getApplication
      responses:
        '200':
          description: Application
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          description: Not found
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
...
  /api/v1/events:
    parameters:
      - $ref: '#/components/parameters/APIVersion'
      - $ref: '#/components/parameters/APIMinVersion'
      - $ref: '#/components/parameters/APIMaxVersion'
    get:
      operationId: streamEvents
      description: Server-sent events stream with stable event IDs and a 24-hour persistent replay window.
      parameters:
        - in: header
          name: Last-Event-ID
          schema:
            type: string
      responses:
        '200':
          description: Event stream
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            text/event-stream:
              schema:
                type: string
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
...
components:
  parameters:
    APIVersion:
      name: Open-Card-API-Version
      in: header
      required: false
      description: >-
        Request an exact API version. This header is mutually exclusive with the minimum and
        maximum version range headers. If all version headers are omitted, API version 1.0 is
        negotiated for legacy clients.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    APIMinVersion:
      name: Open-Card-API-Min-Version
      in: header
      required: false
      description: Lowest API version accepted by the client. Use with Open-Card-API-Max-Version.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    APIMaxVersion:
      name: Open-Card-API-Max-Version
      in: header
      required: false
      description: Highest API version accepted by the client. Use with Open-Card-API-Min-Version.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    ApplicationID:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Stable retry key. Reuse with different input is rejected.
      schema:
        type: string
        minLength: 1
...
  headers:
    APINegotiatedVersion:
      description: API version selected for this request and response.
      schema:
        type: string
        pattern: '^[0-9]+\.[0-9]+$'
  responses:
    APIUpgradeRequired:
      description: No common API version is available for the requested range.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    APIError:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
    Application:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApplicationList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Application'
    CreateApplicationResponse:
      type: object
      required: [application, operation_id]
      properties:
        application:
          $ref: '#/components/schemas/Application'
        operation_id:
          type: string
    CreateApplicationRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
...